test : 사용자가 작성한 시음기록 테스트 이동
- profiles -> records/tasted_record
- url 변경
